+++ src/mmds/src/token.rs
@@ -1,10 +1,6 @@
 // Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-// Temporarily disable unused warnings.
-// TODO: remove these once the Token integration is completed.
-#![allow(dead_code)]
-
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::Read;